move contract verification to its own project
