+++ apps/verify/contract-verify/package.json
@@ -1,0 +1,45 @@
+{
+    "name": "verify",
+    "version": "0.0.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@monaco-editor/react": "^4.5.1",
+        "@next/font": "^13.1.6",
+        "@radix-ui/react-dialog": "^1.0.3",
+        "@radix-ui/react-icons": "^1.3.0",
+        "@uploadcare/upload-client": "^6.6.1",
+        "@vercel/analytics": "^0.1.11",
+        "clsx": "^1.2.1",
+        "dayjs": "^1.11.7",
+        "decimal.js": "^10.4.3",
+        "next": "^13.4.5",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-infinite-scroll-component": "^6.1.0",
+        "react-toastify": "^9.1.3"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.0.0",
+        "@babel/preset-typescript": "^7.21.5",
+        "@types/jest": "^29.5.2",
+        "@types/node": "^17.0.12",
+        "@types/react": "^18.0.22",
+        "@types/react-dom": "^18.0.7",
+        "autoprefixer": "^10.4.13",
+        "eslint": "7.32.0",
+        "eslint-config-custom": "*",
+        "monaco-editor": "^0.38.0",
+        "postcss": "^8.4.21",
+        "prisma": "^4.15.0",
+        "tailwindcss": "^3.2.4",
+        "ts-jest": "^29.1.0",
+        "tsconfig": "*",
+        "typescript": "^4.9.5"
+    }
+}
