+++ apps/verify/contract-verify/pages/_app.tsx
@@ -1,0 +1,8 @@
+import '/workspaces/explorer/apps/web/styles/globals.css';
+import { AppProps } from 'next/app';
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;
