+++ apps/verify/contract-verify/postcss.config.js
@@ -1,0 +1,8 @@
+const config = require("tailwind-config/tailwind.config.js");
+
+module.exports = {
+  plugins: {
+    tailwindcss: { config },
+    autoprefixer: {},
+  },
+}
