+++ apps/web/pages/contract/index.tsx
@@ -1,4 +1,4 @@
-import { VerifyContract } from "../../ui/contract-verify";
+import { VerifyContract } from "../../../verify/contract-verify";
 
 export default function Contract() {
   return (
